+++ fastly/ngwaf/v1/rules/api_get.go
@@ -6,27 +6,33 @@ import (
 	"fmt"
 
 	"github.com/fastly/go-fastly/v10/fastly"
+	"github.com/fastly/go-fastly/v10/fastly/ngwaf/v1/common"
 )
 
-// GetInput specifies the information needed for the Get() function to
-// perform the operation.
+// GetInput specifies the information needed for the Get() function
+// to perform the operation.
 type GetInput struct {
 	// RuleID is the rule identifier (required).
 	RuleID *string
-	// WorkspaceID is the workspace identifier (required).
-	WorkspaceID *string
+	// Scope defines where the rule is applied, including its type
+	// (e.g., "workspace" or "account") and the specific IDs it
+	// applies to (required).
+	Scope *common.Scope
 }
 
 // Get retrieves the specified rule.
 func Get(ctx context.Context, c *fastly.Client, i *GetInput) (*Rule, error) {
-	if i.WorkspaceID == nil {
-		return nil, fastly.ErrMissingWorkspaceID
-	}
 	if i.RuleID == nil {
 		return nil, fastly.ErrMissingRuleID
 	}
+	if i.Scope == nil {
+		return nil, fastly.ErrMissingScope
+	}
 
-	path := fastly.ToSafeURL("ngwaf", "v1", "workspaces", *i.WorkspaceID, "rules", *i.RuleID)
+	path, err := common.BuildPath(i.Scope, "rules", *i.RuleID)
+	if err != nil {
+		return nil, fmt.Errorf("failed to build API path: %w", err)
+	}
 
 	resp, err := c.Get(ctx, path, fastly.CreateRequestOptions())
 	if err != nil {