+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - refactor(ngwaf/v1/workspaces): move all workspace specific endpoints into the appropriate folder ([#708](https://github.com/fastly/go-fastly/pull/708))
 - fix(ngwaf/v1/workpsaces/requests): add missing field to metadata struct and update documentation around limits ([#711](https://github.com/fastly/go-fastly/pull/711))
 - fix(ngwaf/v1/workspaces/redactions): remove mode and page params and remove requirement for both field and type for update ([#712](https://github.com/fastly/go-fastly/pull/712))
+- refactor(ngwaf/v1/rules): combine workspaces and account rules and provide tooling to do the same for lists and signals ([#721](https://github.com/fastly/go-fastly/pull/721))
+
 
 ### Enhancements:
 - feat(ngwaf): add support for alerts ([#714](https://github.com/fastly/go-fastly/pull/714))